Not all tests need to run against the DB
