+++ test/test_data_miner.rb
@@ -1,5 +1,6 @@
 # -*- encoding: utf-8 -*-
 require 'helper'
+init_database
 
 describe DataMiner do
   describe "when used to import example data about pets" do
