+++ test/test_data_miner_run_column_statistic.rb
@@ -1,5 +1,6 @@
 # -*- encoding: utf-8 -*-
 require 'helper'
+init_database
 
 describe DataMiner::Run::ColumnStatistic do
   describe "when advanced statistics are enabled" do
